World management library no longer uses Rectangle2D from the Java standard API
-Added Rect2D class to world package; represents a rectangle in world
space (coordinates at bottom left corner - ascending Y axis)
-Reworked Entity/World2D/WorldImpl classes and all existing demos to use
Rect2D in place of Rectangle2D.

+++ 2DX-GL/demos/com/snap2d/demo/worldgame/ScrollWorld.java
@@ -35,7 +35,7 @@ public class ScrollWorld extends World2D implements GameWorld, Renderable {
 	 * @param ppu
 	 */
 	public ScrollWorld(double minX, double maxY, int viewWidth, int viewHeight,
-			double ppu) {
+			float ppu) {
 		super(minX, maxY, viewWidth, viewHeight, ppu);
 	}
